+++ NEWS.md
@@ -6,6 +6,8 @@
 
 * Added an optional `EXPOSE PORT` line to the generated Dockerfile, to help out Docker Desktop users (#105).
 
+* Added model monitoring dashboard template (#98). To knit the example vetiver monitoring dashboard, execute `vetiver::pin_example_kc_housing_model()` to set up demo model and metrics pins.
+
 # vetiver 0.1.5
 
 * Add functions for model monitoring (#92).